Moved handle_catalog_v2*() methods into a new file
