+++ src/server/internalServer.cpp
@@ -715,79 +715,6 @@ InternalServer::search_catalog(const RequestContext& request,
     return bookIdsToDump;
 }
 
-std::unique_ptr<Response> InternalServer::handle_catalog_v2(const RequestContext& request)
-{
-  if (m_verbose.load()) {
-    printf("** running handle_catalog_v2");
-  }
-
-  std::string url;
-  try {
-    url  = request.get_url_part(2);
-  } catch (const std::out_of_range&) {
-    return Response::build_404(*this, request, "");
-  }
-
-  if (url == "root.xml") {
-    return handle_catalog_v2_root(request);
-  } else if (url == "searchdescription.xml") {
-    const std::string endpoint_root = m_root + "/catalog/v2";
-    return ContentResponse::build(*this,
-        RESOURCE::catalog_v2_searchdescription_xml,
-        kainjow::mustache::object({{"endpoint_root", endpoint_root}}),
-        "application/opensearchdescription+xml"
-    );
-  } else if (url == "entries") {
-    return handle_catalog_v2_entries(request);
-  } else if (url == "categories") {
-    return handle_catalog_v2_categories(request);
-  } else {
-    return Response::build_404(*this, request, "");
-  }
-}
-
-std::unique_ptr<Response> InternalServer::handle_catalog_v2_root(const RequestContext& request)
-{
-  return ContentResponse::build(
-             *this,
-             RESOURCE::catalog_v2_root_xml,
-             kainjow::mustache::object{
-               {"date", gen_date_str()},
-               {"endpoint_root", m_root + "/catalog/v2"},
-               {"feed_id", gen_uuid(m_library_id)},
-               {"all_entries_feed_id", gen_uuid(m_library_id + "/entries")},
-               {"category_list_feed_id", gen_uuid(m_library_id + "/categories")}
-             },
-             "application/atom+xml;profile=opds-catalog;kind=navigation"
-  );
-}
-
-std::unique_ptr<Response> InternalServer::handle_catalog_v2_entries(const RequestContext& request)
-{
-  OPDSDumper opdsDumper(mp_library);
-  opdsDumper.setRootLocation(m_root);
-  opdsDumper.setLibraryId(m_library_id);
-  const auto bookIds = search_catalog(request, opdsDumper);
-  const auto opdsFeed = opdsDumper.dumpOPDSFeedV2(bookIds, request.get_query());
-  return ContentResponse::build(
-             *this,
-             opdsFeed,
-             "application/atom+xml;profile=opds-catalog;kind=acquisition"
-  );
-}
-
-std::unique_ptr<Response> InternalServer::handle_catalog_v2_categories(const RequestContext& request)
-{
-  OPDSDumper opdsDumper(mp_library);
-  opdsDumper.setRootLocation(m_root);
-  opdsDumper.setLibraryId(m_library_id);
-  return ContentResponse::build(
-             *this,
-             opdsDumper.categoriesOPDSFeed(mp_library->getBooksCategories()),
-             "application/atom+xml;profile=opds-catalog;kind=navigation"
-  );
-}
-
 namespace
 {
